test(domain-context): split diagnostic coverage by concern
Constraint: Keep each diagnostic test module below the active 250 pure-LOC threshold.
Rejected: Drop or merge diagnostic assertions | would weaken the locked resolution and privacy coverage
Confidence: high
Scope-risk: narrow
Directive: Keep future diagnostic tests grouped by resolution semantics versus public payload privacy.
Tested: exact 8-test discovery; 215 resolver/privacy/wrapper tests; full Ruff; compileall; git diff --check
Not-tested: full repository unittest discovery
